Create separate directory to traits
